requirements.txt
aaransia==1.1
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.0.5
deep-translator==1.4.2
deepl==0.3
defusedxml==0.7.1
distlib==0.3.1
entrypoints==0.3
facebook-sdk==3.1.0
filelock==3.0.12
Flask==2.0.1
fpdf==1.7.2
futures==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jason==0.1.7
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.5.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
mysql-connector-python==8.0.25
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
notebook==6.3.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2"
prometheus-client==0.10.0
prompt-toolkit==3.0.18
pycparser==2.20
pydeepl==0.13
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.2.1
terminado==0.9.4
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
tweepy==3.10.0
UniCurses==1.2
Unidecode==1.2.0
urllib3==1.26.4
vaderSentiment==3.3.2
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
windows-curses==2.1.0
wordcloud==1.8.1
